+++ src/github.com/matrix-org/dendrite/clientapi/routing/routing.go
@@ -42,6 +42,7 @@ const pathPrefixUnstable = "/_matrix/client/unstable"
 func Setup(
 	servMux *http.ServeMux, httpClient *http.Client, cfg config.Dendrite,
 	producer *producers.RoomserverProducer, queryAPI api.RoomserverQueryAPI,
+	aliasAPI api.RoomserverAliasAPI,
 	accountDB *accounts.Database,
 	deviceDB *devices.Database,
 	federation *gomatrixserverlib.FederationClient,
@@ -109,21 +110,21 @@ func Setup(
 	r0mux.Handle("/directory/room/{roomAlias}",
 		common.MakeAuthAPI("directory_room", deviceDB, func(req *http.Request, device *authtypes.Device) util.JSONResponse {
 			vars := mux.Vars(req)
-			return readers.DirectoryRoom(req, device, vars["roomAlias"], federation, &cfg, queryAPI)
+			return readers.DirectoryRoom(req, device, vars["roomAlias"], federation, &cfg, aliasAPI)
 		}),
 	).Methods("GET")
 
 	r0mux.Handle("/directory/room/{roomAlias}",
 		common.MakeAuthAPI("directory_room", deviceDB, func(req *http.Request, device *authtypes.Device) util.JSONResponse {
 			vars := mux.Vars(req)
-			return readers.SetLocalAlias(req, device, vars["roomAlias"], &cfg, queryAPI)
+			return readers.SetLocalAlias(req, device, vars["roomAlias"], &cfg, aliasAPI)
 		}),
 	).Methods("PUT", "OPTIONS")
 
 	r0mux.Handle("/directory/room/{roomAlias}",
 		common.MakeAuthAPI("directory_room", deviceDB, func(req *http.Request, device *authtypes.Device) util.JSONResponse {
 			vars := mux.Vars(req)
-			return readers.RemoveLocalAlias(req, device, vars["roomAlias"], &cfg, queryAPI)
+			return readers.RemoveLocalAlias(req, device, vars["roomAlias"], &cfg, aliasAPI)
 		}),
 	).Methods("DELETE")
