+++ src/github.com/matrix-org/dendrite/roomserver/storage/storage.go
@@ -355,22 +355,22 @@ func (d *Database) LatestEventIDs(roomNID types.RoomNID) ([]gomatrixserverlib.Ev
 	return references, currentStateSnapshotNID, depth, nil
 }
 
-// SetRoomAlias implements query.RoomserverQueryAPIDB
+// SetRoomAlias implements alias.RoomserverAliasAPIDB
 func (d *Database) SetRoomAlias(alias string, roomID string) error {
 	return d.statements.insertRoomAlias(alias, roomID)
 }
 
-// GetRoomIDFromAlias implements query.RoomserverQueryAPIDB
+// GetRoomIDFromAlias implements alias.RoomserverAliasAPIDB
 func (d *Database) GetRoomIDFromAlias(alias string) (string, error) {
 	return d.statements.selectRoomIDFromAlias(alias)
 }
 
-// GetAliasesFromRoomID implements query.RoomserverQueryAPIDB
+// GetAliasesFromRoomID implements alias.RoomserverAliasAPIDB
 func (d *Database) GetAliasesFromRoomID(roomID string) ([]string, error) {
 	return d.statements.selectAliasesFromRoomID(roomID)
 }
 
-// RemoveRoomAlias implements query.RoomserverQueryAPIDB
+// RemoveRoomAlias implements alias.RoomserverAliasAPIDB
 func (d *Database) RemoveRoomAlias(alias string) error {
 	return d.statements.deleteRoomAlias(alias)
 }
